APN complete
APN for regular notifications
APN for follow up notifications
App checks to see if local data valid when trying to show alarm
APN makes loud notification sound ( if desired )
Added refresh buttons to Logs and Reminders page
Revised Personal Info and Family settings pages to have more options + be more clean.

+++ Hound Specific/Server/Node/controllers/deleteFor/deleteForUser.js
@@ -2,6 +2,7 @@ const DatabaseError = require('../../utils/errors/databaseError');
 const ValidationError = require('../../utils/errors/validationError');
 const { queryPromise } = require('../../utils/database/queryPromise');
 const { deleteDogsQuery } = require('./deleteForDogs');
+const { deleteSecondaryAlarmNotificationsForUser } = require('../../utils/notification/alarm/deleteAlarmNotification');
 
 /**
  *  Queries the database to delete a user and everything nested under it. If the query is successful, then returns
@@ -28,44 +29,33 @@ const deleteUserQuery = async (req, userId, familyId) => {
     throw new DatabaseError(error.code);
   }
 
-  // User is not the head of their family and has no responsibility. Can leave family and delete user
-  if (familyHeads.length === 0) {
+  // Either:
+  // 1. user not familyHead, so they can leave family and we can delete the user
+  // 2. user is familyHead but only familyMember, so we can delete the family and delete the user
+  if (familyHeads.length === 0 || (familyHeads.length === 1 && familyMembers.length === 1)) {
     try {
-      // deletes user from family
-      await queryPromise(req, 'DELETE FROM familyMembers WHERE userId = ?', [userId]);
+      if (familyHeads.length === 1 && familyMembers.length === 1) {
+        // User is the ONLY family member and the familyHead. We will delete the family
 
-      // perform rest of deletion
-
-      // delete userConfiguration
-      await queryPromise(req, 'DELETE FROM userConfiguration WHERE userId = ?', [userId]);
-      // deletes user
-      await queryPromise(req, 'DELETE FROM users WHERE userId = ?', [userId]);
-    }
-    catch (error) {
-      throw new DatabaseError(error.code);
-    }
-  }
-  // User is the head of their family and is the ONLY family member. Therefore no responsibility so can delete family and delete user
-  else if (familyMembers.length === 1) {
-    try {
-      // delete the family head which is the user
-      await queryPromise(req, 'DELETE FROM familyHeads WHERE userId = ?', [userId]);
-
-      // perform rest of deletion
-
+        // delete the family head which is the user
+        await queryPromise(req, 'DELETE FROM familyHeads WHERE userId = ?', [userId]);
+        // delete all the dogs in the family since deleting the family
+        await deleteDogsQuery(req, userId, familyId);
+      }
+      // No matter what, we delete the user and their information
       // deletes user from family
       await queryPromise(req, 'DELETE FROM familyMembers WHERE userId = ?', [userId]);
       // delete userConfiguration
       await queryPromise(req, 'DELETE FROM userConfiguration WHERE userId = ?', [userId]);
       // deletes user
       await queryPromise(req, 'DELETE FROM users WHERE userId = ?', [userId]);
+      // delete their secondary notifications
+      // technically not required to do as their notification token will be deleted (so no APN will be sent), but this reduces the load on our schedules
+      deleteSecondaryAlarmNotificationsForUser(userId);
     }
     catch (error) {
       throw new DatabaseError(error.code);
     }
-
-    // delete all the dogs
-    await deleteDogsQuery(req, familyId);
   }
   else {
     throw new ValidationError("User's family still has other members in it", 'ER_VALUES_INVALID');
